Add File Recognition and improve SQl save function. Now you can set database and table names.

+++ Converter/mainwindow.cpp
@@ -11,9 +11,16 @@
 #include "QSqlQuery"
 #include "QString"
 #include <qsqlrecord.h>
+#include <QFileInfo>
+#include <QDebug>
+#include <QSqlError>
+#include "QInputDialog";
 
 QString FileWay="";
 QSqlDatabase db;
+QString MainDataType;
+
+
 
 MainWindow::MainWindow(QWidget *parent) :
     QMainWindow(parent),
@@ -35,42 +42,56 @@ void MainWindow::on_OpenButton_clicked()
     QStringList horizontalHeader;
     QStringList TableItem;
     QStringList RowItem;
+
+//    QString FileWay=QFileDialog::getOpenFileName(this,tr("Open CSV or Sqlite file"),"/home/bi",tr("CSV files (*.csv);;All files (*.*)"));
     QString FileWay=QFileDialog::getOpenFileName();
     ui->FileWaylabel->setText(FileWay);
+
+
+
     QFile file(FileWay);
-    if(!file.open(QIODevice::ReadOnly))
-        QMessageBox::information(0,"info",file.errorString());
-        QTextStream Data(&file);
+     if (FileWay.contains(".csv")){
+         MainDataType="CSV";
 
-        QString text= Data.readAll();
-        file.close();
-        RowItem=text.split("\n");
-        for(int i=0;i<RowItem.size()-1;i++)
-        {
-            QString NewString=RowItem.at(i);
-            NewString.remove(0,1);
-            NewString.remove(NewString.size()-1,1);
-            TableItem=NewString.split("\";\"");
+        if(!file.open(QIODevice::ReadOnly))
+            QMessageBox::information(0,"info",file.errorString());
+            QTextStream Data(&file);
 
-            for(int j=0; j<TableItem.size();j++)
+            QString text= Data.readAll();
+            file.close();
+            RowItem=text.split("\n");
+            for(int i=0;i<RowItem.size()-1;i++)
             {
+                QString NewString=RowItem.at(i);
+                NewString.remove(0,1);
+                NewString.remove(NewString.size()-1,1);
+                TableItem=NewString.split("\";\"");
 
-                if (i==0){
+                for(int j=0; j<TableItem.size();j++)
+                {
 
-                    horizontalHeader.append(TableItem[j]);
+                    if (i==0){
 
-                }else{
-                    item = new QStandardItem(TableItem[j]);
-                    model->setItem(i-1, j, item);
+                        horizontalHeader.append(TableItem[j]);
+
+                    }else{
+                        item = new QStandardItem(TableItem[j]);
+                        model->setItem(i-1, j, item);
+                    }
                 }
             }
-        }
 
-            model->setHorizontalHeaderLabels(horizontalHeader);
-            ui->tableView->setModel(model);
+                model->setHorizontalHeaderLabels(horizontalHeader);
+                ui->tableView->setModel(model);
 
-            ui->tableView->resizeRowsToContents();
-            ui->tableView->resizeColumnsToContents();
+                ui->tableView->resizeRowsToContents();
+                ui->tableView->resizeColumnsToContents();
+     }
+     else
+     {
+         MainDataType="SQLite";
+     }
+
 
 }
 
@@ -82,64 +103,86 @@ void MainWindow::on_ConvertButton_clicked()
     QString TextValue;
     QString AllHeaders;
     QString StringData;
-    //QString SomeString = ui->tableView->model()->data(ui->tableView->model()->index(0, 0)).toString();
-    FileWay=QFileDialog::getOpenFileName();
-    db=QSqlDatabase::addDatabase("QSQLITE");
-    db.setDatabaseName(FileWay);
-    db.open();
+    bool NewTableButtonOk;
+    if(ui->FileWaylabel->text()!=""){
+        if(MainDataType=="CSV"){
+            FileWay=QFileDialog::getSaveFileName();
+                db=QSqlDatabase::addDatabase("QSQLITE");
+                db.setDatabaseName(FileWay);
+                db.open();
+                if(db.open())
+                {
+                    QString TableName=QInputDialog::getText(0,"New Table","Table Name",QLineEdit::Normal,"",&NewTableButtonOk);
+                    if(NewTableButtonOk){
+
+                        for( int i = 0; i < ui->tableView->horizontalHeader()->count(); ++i )
+                        {
+                            QString HeaderNames = ui->tableView->model()->headerData(i, Qt::Horizontal).toString();
+                            SqlQueryString+=" "+HeaderNames;
+                            AllHeaders+=" "+HeaderNames+",";
+                            for( int j = 0; j < ui->tableView->verticalHeader()->count(); ++j )
+                            {
+                              TextValue = ui->tableView->model()->data(ui->tableView->model()->index(j, i)).toString();
 
 
-    QString Fname="NewSUperDB";
-    for( int i = 0; i < ui->tableView->horizontalHeader()->count(); ++i )
-        {
-            QString HeaderNames = ui->tableView->model()->headerData(i, Qt::Horizontal).toString();
-            SqlQueryString+=" "+HeaderNames;
-            AllHeaders+=" "+HeaderNames+",";
-            for( int j = 0; j < ui->tableView->verticalHeader()->count(); ++j )
-            {
-              TextValue = ui->tableView->model()->data(ui->tableView->model()->index(j, i)).toString();
+                              bool success;
+                              int ConvertIntData =TextValue.toInt(&success);
+                              if((success)&&(DataType=="INTEGER"))
+                              {
+                                  DataType="INTEGER";
+                              }
+                              else
+                              {
+                                  double ConvetDoubleData=TextValue.toDouble(&success);
+                                  if((success))
+                                  {
+                                      DataType="REAL";
+                                  }
+                                  else
+                                  {
+                                  DataType="TEXT";
+                                  break;
+                                  }
+                               }
+                              }
+                            SqlQueryString+=" "+DataType+" ,";
+
+                            DataType="INTEGER";
+                            }
+                    SqlQueryString.remove(SqlQueryString.size()-1,1);
+                    AllHeaders.remove(AllHeaders.size()-1,1);
 
 
-              bool success;
-              int ConvertData =TextValue.toInt(&success);
-              if((success)&&(DataType=="INTEGER")){
-                  DataType="INTEGER";
-              } else{
-                  DataType="TEXT";
-                 }
-              }
-            SqlQueryString+=" "+DataType+" ,";
+                    QSqlQuery query;
+                    query.prepare("CREATE TABLE "+TableName+"("+SqlQueryString+");");
 
-            DataType="INTEGER";
+                    query.exec();
+                    for( int j = 0; j < ui->tableView->verticalHeader()->count(); ++j )
+                    {
+                        for( int i = 0; i < ui->tableView->horizontalHeader()->count(); ++i )
+                        {
+                            TextValue = ui->tableView->model()->data(ui->tableView->model()->index(j, i)).toString();
+                            StringData+="'"+TextValue+"' ,";
+                        }
+                        StringData.remove(StringData.size()-1,1);
+                        //ui->FileWaylabel->setText("INSERT INTO "+Fname+"("+AllHeaders+") VALUES("+StringData+");");
+                        query.prepare("INSERT INTO "+TableName+"("+AllHeaders+") VALUES("+StringData+");");
+                        query.exec();
+                        StringData="";
+
+                    }
+
+                    db.close();
+                    }
             }
-    SqlQueryString.remove(SqlQueryString.size()-1,1);
-    AllHeaders.remove(AllHeaders.size()-1,1);
-
-
-    QSqlQuery query;
-    query.prepare("CREATE TABLE "+Fname+"("+SqlQueryString+");");
-    query.exec();
-    for( int j = 0; j < ui->tableView->verticalHeader()->count(); ++j )
-    {
-        for( int i = 0; i < ui->tableView->horizontalHeader()->count(); ++i )
-        {
-            TextValue = ui->tableView->model()->data(ui->tableView->model()->index(j, i)).toString();
-            StringData+="'"+TextValue+"' ,";
-
-
-
+            else
+            {
+                qDebug()<<db.lastError().text();
+            }
         }
-        StringData.remove(StringData.size()-1,1);
-        ui->FileWaylabel->setText("INSERT INTO "+Fname+"("+AllHeaders+") VALUES("+StringData+");");
-        query.prepare("INSERT INTO "+Fname+"("+AllHeaders+") VALUES("+StringData+");");
-        query.exec();
-        StringData="";
+   }
 
-    }
-
-    db.close();
-
-    }
+}
